+++ src/gr_logging.cpp
@@ -213,7 +213,7 @@ json & graphite2::operator << (json & j, const dslot & ds) throw()
             << "offset"         << s.attachOffset()
             << json::close;
     j << "user" << json::flat << json::array;
-    for (int n = 0; n!= seg.numAttrs(); ++n)
+    for (size_t n = 0; n != seg.numAttrs(); ++n)
         j   << s.userAttrs()[n];
     j       << json::close;
     if (s.firstChild())
