+++ src/inc/Silf.h
@@ -79,7 +79,7 @@ public:
     uint16 findClassIndex(uint16 cid, uint16 gid) const;
     uint16 getClassGlyph(uint16 cid, unsigned int index) const;
     uint16 findPseudo(uint32 uid) const;
-    uint8 numUser() const { return m_aUser; }
+    size_t numUser() const { return m_aUser; }
     uint8 aPseudo() const { return m_aPseudo; }
     uint8 aBreak() const { return m_aBreak; }
     uint8 aMirror() const {return m_aMirror; }
@@ -90,9 +90,9 @@ public:
     uint8 positionPass() const { return m_pPass; }
     uint8 justificationPass() const { return m_jPass; }
     uint8 bidiPass() const { return m_bPass; }
-    uint8 numPasses() const { return m_numPasses; }
+    size_t numPasses() const { return m_numPasses; }
     uint8 maxCompPerLig() const { return m_iMaxComp; }
-    uint16 numClasses() const { return m_nClass; }
+    size_t numClasses() const { return m_nClass; }
     byte  flags() const { return m_flags; }
     byte  dir() const { return m_dir; }
     size_t numJustLevels() const { return m_numJusts; }
